pyproject.toml
[project]
name = "amazon-sales-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "mlflow>=3.7.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "streamlit>=1.52.1",
    "uvicorn[standard]>=0.38.0",
    "xgboost>=3.1.2",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.0",
    "pytest>=9.0.2",
]
